ya funcionan las pruebas

+++ src/main/java/com/teamcreator/creator/Repositorios/UsersPorEquipoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.teamcreator.creator.Modelo.UsersPorEquipo;
 
-public interface UsersPorEquipoRepository extends JpaRepository<UsersPorEquipo, Long> {
+public interface UsersPorEquipoRepository extends JpaRepository<UsersPorEquipo, Integer> {
 
     List<UsersPorEquipo> findByEquipos(String nombre_equipo);
